requirements.txt
absl-py==2.1.0
astunparse==1.6.3
capstone==5.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.6.0
keystone-engine==0.9.2
kiwisolver==1.4.7
libclang==18.1.1
lief==0.15.1
lightgbm==4.5.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.5.82
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pillow==11.0.0
protobuf==5.28.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
r2pipe==1.9.4
ratelimit==2.2.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
six==1.16.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.2
wheel==0.44.0
wrapt==1.16.0
